Version 8.3.7
Fixed some lowercase/uppercase bugs in ranks. (permissions still need to
be lowercase)

Fixed the way treasures are updated. It's the way it was, before fake
inventories were introduced.
Just use /tchest set, to update the items of a treasure. The plugin will
use the fake inventory, if that makes sense.
You no longer need to click the chest with a chest, it was too
confusing.