Add group support
This commit adds support for creating and managing groups of chests. A
group can have any number of chests in them (as long as all chests are
in the same world) and a chest can belong to any number of groups.

New commands added:
group-create :- Create a group to add chests to (in the world your in)
group-delete :- Delete the specified group
group-add :- Add the chest your looking at to the specified group
group-remove :- Remove the chest your looking at from the specified
group
group-forget :- Tell all chests in the specified group to forget a
specific player
group-forget-all :- Tell all chests in the specified group to forget all
players
group-copy :- Copy the contents of the chest your looking at to all
chests in the specified group