Add support for single inventory chests
Via the use of setSingle a treasure chest can be created which rather
then having a per player inventory has a single inventory. The addition
of this feature allows treasurechest to be used in group games where you
want to restrict some chests to "first come first served".

Note: The life-time of the inventory for this type of chest is defined
by either server reset, or the use of the forget command.