Fix glaring flaw in singleton chests
A fully raided chest as the same inventory as a new inventory created by
createTreasureInventory so don't try and detect a new chest inventory
being created that way.

Created a 2nd createTreasureInventory function which can be used to
detect if a chest  inventory already exists for the specified input,
using this we can determine if this is the first time a singleton chest
has been opened.