Add region commands
Add the ability to select a region (currently requires WE support) and then run commands which will effect the selected container blocks. The aim is to allow quicker and more comprehensive set-up features especially when working on arenas built up via WE etc . The new commands are as follows:

/tchest region set [block,white-list] Create/update all treasure in the selected region
/tchest region set-shared [block,white-list] Create/update all treasure with a shared inventory in the selected region
/tchest region group-add <group name> [block,white-list] Add all treasure in the selected region to a group
/tchest region group-remove <group name> [block,white-list] Remove all treasure in the selected region from a group
/tchest region use-meta-data [block,white-list] Create/update all treasure in the selected region using the meta-data of items in the chest (sticks with names where shared will make a chest shared and any other name will add the chest to a group with that name)
/tchest region delete [block,white-list] Delete all treasure in the selected region

As you can see all these commands take an optional block white-list, if this is not provided then the command will work on all the container blocks in the region and if it is then it will only work on the list provided which should be comma separated and without spaces.